rename 'symbols' and 'rel_symbols' to be clearer, and remove redundant overrides of keep_section
